requirements.txt
flask>=2.0.0
tensorflow>=2.10.0
Pillow>=9.0.0
numpy>=1.21.0
werkzeug>=2.0.0

